fix( routing ): remove improper multi-step carousel functionality for signup form
This commit removes the improperly implemented multi-step carousel
functionality from the Signup route. The carousel code was removed because it
was leading to confusion and potential bugs. This change simplifies the
Signup form and improves the overall user experience since instead of using a
library for multi-step forms, we can implement a custom solution later.