feat( auth ): add AuthProvider for handling global auth state on Frontend and it's associated components
This commit introduces the AuthProvider component to manage user authentication
state across the application. It includes methods for signing in and signing out
users, as well as maintaining the current user's information. The Signin route
has been updated to utilize the AuthProvider for handling user sign-in, and a
ProtectedRoute component has been added to restrict access to certain routes
based on authentication status. Additionally, the NavAvatar component now
includes a sign-out option that leverages the AuthProvider's sign-out
functionality.