fix( auth ): update auth provider and related components to improve error handling
This commit enhances the error handling mechanisms in the AuthProvider as well as
in the Signin route, Signup route and the NavAvatar component. Instead of returning
generic error messages, the code now returns more specific error information received
from the backend API. This allows for better user feedback and debugging.